Polish fixed-size binary implementation
